pyproject.toml
[project]
name = "dual-llm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.128.0",
    "google-generativeai>=0.8.6",
    "huggingface-hub>=1.3.2",
    "langchain>=1.2.6",
    "langchain-google-genai>=4.2.0",
    "langchain-groq>=1.1.1",
    "langchain-openai>=1.1.7",
    "python-dotenv>=1.2.1",
    "streamlit>=1.53.0",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "black>=25.12.0",
    "ipykernel>=7.1.0",
    "ruff>=0.14.11",
]
